pyproject.toml
[tool.poetry]
name = "resume-generator"
version = "0.1.0"
description = ""
authors = ["nerdimite <bhavesh.laddagiri1@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.57.3"
pymupdf = "^1.25.1"
tenacity = "^9.0.0"
pydantic = "^2.10.3"
jinja2 = "^3.1.4"
pdfkit = "^1.0.0"
python-dotenv = "^1.0.1"
pytz = "^2024.2"
numpy = "^2.2.0"
pillow = "^11.0.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
